edupilot-backend/requirements.txt
asgiref==3.7.2
certifi==2025.4.26
charset-normalizer==3.4.2
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.10
numpy==1.26.4
pandas==2.2.2
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2023.3.post1
requests==2.32.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.4.0
